requirements.txt
flask
flask_uploads
flask_dropzone
Pillow
tensorflow
